+++ src/helpers/radCli.ts
@@ -1,6 +1,4 @@
-import type { Patch } from '../types'
-import { useEnvStore } from '../stores'
-import { assertUnreachable, isRealFsPath, log } from '../utils'
+import { assertUnreachable, isRealFsPath } from '../utils'
 import { exec, execRad, getResolvedPathToNodeHome } from '.'
 
 /**
@@ -176,42 +174,4 @@ export function getNodeSshKey(format: 'fingerprint' | 'full'): string | undefine
   const { stdout: nodeSshKey, errorCode } = execRad(['self', flag])
 
   return errorCode ? undefined : nodeSshKey
-}
-
-/**
- * Performs edition of the patch's title and description on Radicle. Effectively
- * it's editing the first patch revision.
- *
- * @returns An object with the outcome of the editing operation plus additional details
- * when available
- */
-export function editPatch(
-  patchId: Patch['id'],
-  newTitle: string,
-  newDescr: string,
-  timeoutSeconds?: number,
-): { outcome: 'success'; didAnnounce: boolean } | { outcome: 'failure'; errorMsg: string } {
-  const rid = useEnvStore().currentRepoId
-  if (!rid) {
-    const errorMsg = 'Unable to resolve current repo id in `updatePatchTitleAndDescription()`'
-    log(errorMsg, 'error')
-
-    return { outcome: 'failure', errorMsg }
-  }
-
-  const execResult = execRad(
-    ['patch', 'edit', patchId, '--repo', rid, '--message', newTitle, '--message', newDescr],
-    { shouldLog: true, timeout: timeoutSeconds ? timeoutSeconds * 1000 : undefined },
-  )
-
-  if (execResult.errorCode) {
-    return {
-      outcome: 'failure',
-      errorMsg: `${execResult.stdout}\n${execResult.stderr}\n${execResult.errorCode}`,
-    }
-  } else if (execResult.stdout?.includes('Node is stopped')) {
-    return { outcome: 'success', didAnnounce: false }
-  } else {
-    return { outcome: 'success', didAnnounce: true }
-  }
 }